requirements.txt
absl-py==2.1.0
actionlib==1.14.0
addict==2.4.0
albumentations==1.3.0
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
angles==1.9.13
arbotix-python==0.11.0
baldor==0.1.3
base_local_planner==1.17.3
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cob-cartesian-controller==0.8.22
cob-script-server==0.6.33
cob-twist-controller==0.8.22
colorama==0.4.6
contourpy==1.2.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
controller-manager-tests==0.20.0
crcmod==1.7
cryptography==42.0.5
cv-bridge==1.16.2
cycler==0.12.1
diagnostic-updater==1.11.0
dynamic-reconfigure==1.7.3
dynamixel_sdk==3.7.51
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.3.1
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
grpcio==1.62.1
h5py==3.10.0
idna==3.6
image-geometry==1.16.2
imageio==2.34.0
imgaug==0.4.0
importlib_metadata==7.1.0
interactive-markers==1.12.0
Jinja2==3.1.3
jmespath==0.10.0
joblib==1.3.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
kiwisolver==1.4.5
laser_geometry==1.6.7
lazy_loader==0.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
message-filters==1.16.0
mmcv-full==1.7.2
model-index==0.1.11
moveit-commander==1.1.13
moveit-core==1.1.13
moveit-python==0.4.5
moveit-ros-planning-interface==1.1.13
moveit-ros-visualization==1.1.13
moveit-task-constructor-core==0.1.3
mpmath==1.3.0
natsort==8.4.0
networkx==3.2.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.36
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
pr2_controller_manager==1.8.21
protobuf==5.26.0
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
pytz==2023.4
PyYAML==6.0.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qudida==0.0.4
requests==2.28.2
resource_retriever==1.12.7
rich==13.4.2
rosbag==1.16.0
rosclean==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.16.0
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-console==0.4.12
rqt-controller-manager==0.20.0
rqt-image-view==0.4.17
rqt-joint-trajectory-controller==0.21.2
rqt-logger-level==0.4.12
rqt-reconfigure==0.5.5
rqt-tf-tree==0.6.4
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_dep==0.4.12
rqt_graph==0.4.14
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_launch==0.4.9
rqt_msg==0.4.10
rqt_plot==0.4.13
rqt_publisher==0.4.10
rqt_py_common==0.5.3
rqt_py_console==0.4.10
rqt_service_caller==0.4.10
rqt_shell==0.4.11
rqt_srv==0.4.9
rqt_top==0.4.10
rqt_topic==0.4.13
rqt_web==0.4.10
rviz==1.14.20
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
sensor-msgs==1.13.1
shapely==2.0.3
six==1.16.0
smclib==1.8.6
srdfdom==0.6.4
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.2.12
tomli==2.0.1
topic-tools==1.16.0
torch==2.1.2
torchaudio==2.1.2
torchsummary==1.5.1
torchvision==0.16.2
tqdm==4.65.2
triton==2.1.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.9.0
urdfdom-py==0.4.6
urllib3==1.26.18
Werkzeug==3.0.1
xacro==1.14.16
yapf==0.40.2
zipp==3.18.1
